+++ plugin/src/types.ts
@@ -1,0 +1,8 @@
+export interface BitmovinConvivaAnalyticsPluginProps {
+  /**
+   * Version of BitmovinConvivaAnalytics iOS pod to use
+   * Must be >= 3.6.1
+   * @default "3.6.1"
+   */
+  bitmovinConvivaIosVersion?: string;
+}